requirements.txt
keras==2.12.0
keras==2.14.0rc0
keras==2.10.0
keras==2.11.0
keras_nightly==2.5.0.dev2021032900
numpy==1.25.1
numpy==1.19.5
numpy==1.15.1
numpy==1.21.6
numpy==1.21.5
numpy==1.23.5
numpy==1.25.2
numpy==1.24.3
opencv_contrib_python==4.5.5.64
opencv_python==4.5.5.64
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.1.0
Pillow==10.0.1
Pillow==10.0.0
Pillow==9.5.0
Pillow==9.0.1
Pillow==9.3.0
Pillow==9.4.0
Pillow==10.0.1
streamlit==1.23.1
streamlit==1.22.0
streamlit==1.26.0
tensorflow==2.12.0
tensorflow==2.13.0
tensorflow==2.14.0rc0
tensorflow==2.10.0
tensorflow==2.6.0
tensorflow==2.11.0
tensorflow_intel==2.12.0
tensorflow_intel==2.13.0
tensorflow_intel==2.14.0rc0
tensorflow_intel==2.11.0
torch==1.13.1
torch==2.0.1
~treamlit==1.22.0
